4.x: Add ZeroTokenNodesIT (#402)
* CcmBridge methods adjustments

Adds `startWithArgs` methods that allow avoiding the default arguments like
"--wait-other-notice". This wait causes timeout in multi-dc test
utilizing `join_ring: false` so it needs to be avoided.

Adds `addWithoutStart(n, dc)`. As the name suggests it's an add that does not
immediately start the node.

Adds `updateNodeConfig` in order to allow adding new nodes with different
configurations.

* Add ZeroTokenNodesIT

Adds a test class verifying the behavior of the driver when dealing with
zero-token Scylla nodes.

Adds a configurable option which controls whether to allow zero-token nodes as
valid peers. Until now such nodes were considered invalid. Default value of
this option does not change the driver behavior.