4.x: Adjust `BasicLoadBalancingPolicy#getReplicas` (#405)
* Fetch additional keyspace metadata from scylla_keyspaces

Extends schema related classes with methods for querying scylla_keyspaces
in order to determine whether keyspace is tablets-enabled.

Adjusts schema queries tests and several other integration tests
that mock responses from system tables.

Adds `isUsingTablets()` and `setUsingTablets()` to KeyspaceMetadata.

* Use keyspace metadata in `BasicLoadBalancingPolicy#getReplicas`

Modifies the getReplicas method to do the tablet lookup if and only if the
keyspace metadata indicates that it's a tablets-based keyspace. Otherwise refer
to the token map.
Previous behavior was to try tablet map lookup first regardless of the
keyspace configuration.