Allow arbitrary Scylla versions in CcmBridge
Until now, the versions passed through ccm.version property would be
considered invalid if cannot be parsed to one of the known formats.
Now if the parsing fails, the property will be passed to ccm commands
"as is" which ccm will treat as directory name in local repository.
Numbered version will then be fetched through setting up a makeshift cluster
and asking ccm for it.